final dto

+++ src/main/java/org/iesvdm/modelo/Cliente.java
@@ -5,15 +5,20 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+//La anotación @Data de lombok proporcionará el código de:
+//getters/setters, toString, equals y hashCode
+//propio de los objetos POJOS o tipo Beans
 @Data
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
 public class Cliente {
-	private long id;
+	
+	private int id;
 	private String nombre;
 	private String apellido1;
 	private String apellido2;
 	private String ciudad;
 	private int categoria;
+	
 }
